pyproject.toml
[tool.poetry]
name = "dbt-airflow"
version = "0.3.2"
description = "A Python package that creates fine-grained Airflow tasks for dbt"
repository = "https://github.com/gmyrianthous/dbt-airflow"
authors = ["Giorgos Myrianthous <giorgos.myrianthous@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "dbt_airflow" }]

[tool.poetry.dependencies]
python = ">=3.7.2,<4"
dbt-core = ">=1.2.0"
apache-airflow = ">=2.0.0,<3"
pydantic = "^1.10.4"
apache-airflow-providers-cncf-kubernetes = ">=6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.1,<8"
commitizen = "^2.40.0"
mkdocs-material = "^8.2"
mkdocstrings-python = "^0.8.2"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.3.2"
version_files = [
    "pyproject.toml:^version"
]
update_changelog_on_bump = true
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
